azure-function-hhh/requirements.txt
azure-functions
pydantic
anthropic